Reverted changes to revit wall subtypes
